requirements.txt
django==3.1
djangorestframework==3.12.4
django-filter==2.4.0
